better keyboard tracking for GridSplitter
